fix: corrige bugs de user-svc detectados al probar con Docker
- Repositorios (Usuario, Huesped, Propietario, CuentaBancaria, TarjetaCredito) declaraban ID Long pero las entidades usan Integer: nunca compilaba.
- ControllerAdvisor devolvia 500 en vez de 405 para metodos HTTP no soportados.
- HuespedRecord.toHuesped() armaba una TarjetaCredito embebida sin el back-reference al huesped, rompiendo el cascade al guardar.
- crearUsuarioHuesped/crearUsuarioPropietario sin @Transactional dejaban usuarios huerfanos si fallaba el paso siguiente.
- CuentaBancaria y Propietario se serializaban en un ciclo infinito (faltaba @JsonIgnore en el lado mappedBy).
- HuespedUpdateDTO/PropietarioUpdateDTO exigian nombre con @NotBlank, rompiendo las actualizaciones parciales que ya soporta el service.
- Columna fecha_vencimiento era VARCHAR(4), insuficiente para el formato MM/YY validado en el DTO.
- Postman de ETAPA01 no coincidia con la forma real de los DTOs (Crear Huesped anidaba tarjetaCredito, Crear Propietario le faltaba alias).

+++ services/user-svc/src/main/java/edu/utn/frsf/isi/dan/user/dao/TarjetaCreditoRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Long> {
+public interface TarjetaCreditoRepository extends JpaRepository<TarjetaCredito, Integer> {
     
     Optional<TarjetaCredito> findByHuespedIdAndEsPrincipalTrue(Integer huespedId);
